Move unit test to a new file
