+++ tests/loggers/test_base.py
@@ -176,74 +176,6 @@ def test_multiple_loggers(tmpdir):
     assert logger2.finalized_status == "success"
 
 
-def test_trainer_loggers_property():
-    logger1 = CustomLogger()
-    logger2 = CustomLogger()
-
-    # trainer.loggers should be a copy of the input list
-    trainer = Trainer(logger=[logger1, logger2])
-
-    assert trainer.loggers == [logger1, logger2]
-
-    # trainer.loggers should create a list of size 1
-    trainer = Trainer(logger=logger1)
-
-    assert trainer.loggers == [logger1]
-
-    # trainer.loggers should be an empty list
-    trainer = Trainer(logger=False)
-
-    assert trainer.loggers == []
-
-    # trainer.loggers should be a list of size 1 holding the default logger
-    trainer = Trainer(logger=True)
-
-    assert trainer.loggers == [trainer.logger]
-    assert type(trainer.loggers[0]) == TensorBoardLogger
-
-
-def test_trainer_logger_setters():
-    logger1 = CustomLogger()
-    logger2 = CustomLogger()
-    logger_collection = LoggerCollection([logger1, logger2])
-
-    trainer = Trainer()
-    assert type(trainer.logger) == TensorBoardLogger
-    assert trainer.loggers == [trainer.logger]
-
-    # Test setters for trainer.logger
-    trainer.logger = logger1
-    assert trainer.logger == logger1
-    assert trainer.loggers == [logger1]
-
-    trainer.logger = logger_collection
-    assert trainer.logger._logger_iterable == logger_collection._logger_iterable
-    assert trainer.loggers == [logger1, logger2]
-
-    trainer.logger = None
-    assert trainer.logger is None
-    assert trainer.loggers == []
-
-    # Test setters for trainer.loggers
-    trainer.loggers = [logger1, logger2]
-    assert trainer.loggers == [logger1, logger2]
-    assert trainer.logger._logger_iterable == logger_collection._logger_iterable
-
-    trainer.loggers = [logger1]
-    assert trainer.loggers == [logger1]
-    assert trainer.logger == logger1
-
-    trainer.loggers = logger_collection
-    assert trainer.loggers == [logger1, logger2]
-    assert trainer.logger._logger_iterable == logger_collection._logger_iterable
-
-    trainer.loggers = []
-    assert trainer.loggers == []
-    assert trainer.logger is None
-
-    trainer.loggers = None
-    assert trainer.loggers == []
-    assert trainer.logger is None
 
 
 def test_multiple_loggers_pickle(tmpdir):
